重构 AdaptiveComicCover，详情页封面固定比例白底 contain 并缩小圆角。
