+++ app/lib/ui/features/library/views/desktop/comic_detail_page/widgets/comic_detail_cover.dart
@@ -3,7 +3,7 @@ import 'package:hentai_library/ui/core/theme/theme.dart';
 import 'package:hentai_library/ui/core/dto/comic_cover_display_data.dart';
 import 'package:hentai_library/src/rust/api/thumbnail.dart';
 import 'package:hentai_library/ui/providers.dart';
-import 'package:hentai_library/ui/core/widgets/element/image/adaptive_cover.dart';
+import 'package:hentai_library/ui/core/widgets/element/image/adaptive_comic_cover.dart';
 import 'package:hentai_library/domain/models/entity/comic/comic.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
@@ -13,10 +13,12 @@ class ComicDetailCover extends ConsumerWidget {
 
   final Comic comic;
 
+  static const double containerAspectRatio = 2 / 3;
+  static const double _cornerRadius = 2;
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final ColorScheme cs = Theme.of(context).colorScheme;
-    final AppThemeTokens tokens = context.tokens;
     final ComicCoverDisplayData? coverData = ref
         .watch(
           comicCoverDisplayProvider(
@@ -26,10 +28,11 @@ class ComicDetailCover extends ConsumerWidget {
         )
         .maybeWhen(data: (ComicCoverDisplayData? v) => v, orElse: () => null);
 
-    return AdaptiveCover(
+    return AdaptiveComicCover(
       coverDisplay: coverData,
-      fallbackAspectRatio: 2 / 3,
-      backgroundColor: cs.hentai.imagePlaceholder,
+      containerAspectRatio: containerAspectRatio,
+      backgroundColor: Colors.white,
+      showShadow: true,
       placeholder: const SizedBox.expand(),
       errorPlaceholder: Center(
         child: Icon(
@@ -38,7 +41,7 @@ class ComicDetailCover extends ConsumerWidget {
           size: 40,
         ),
       ),
-      clipBorderRadius: BorderRadius.circular(tokens.radius.lg),
+      clipBorderRadius: BorderRadius.circular(_cornerRadius),
     );
   }
 }
