+++ app/lib/ui/features/library/views/desktop/series_detail_page/widgets/series_detail_cover.dart
@@ -4,7 +4,7 @@ import 'package:hentai_library/domain/models/entity/comic/series.dart';
 import 'package:hentai_library/domain/models/entity/comic/series_item.dart';
 import 'package:hentai_library/ui/core/dto/comic_cover_display_data.dart';
 import 'package:hentai_library/ui/providers.dart';
-import 'package:hentai_library/ui/core/widgets/element/image/adaptive_cover.dart';
+import 'package:hentai_library/ui/core/widgets/element/image/adaptive_comic_cover.dart';
 import 'package:hentai_library/ui/core/theme/theme.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 
@@ -29,7 +29,7 @@ class SeriesDetailCover extends ConsumerWidget {
     final AppThemeTokens tokens = context.tokens;
     final ColorScheme cs = Theme.of(context).colorScheme;
 
-    return AdaptiveCover(
+    return AdaptiveComicCover(
       coverDisplay: coverDisplay,
       fallbackAspectRatio: 2 / 3,
       backgroundColor: cs.hentai.imagePlaceholder,
